ai-decision-intelligence-platform/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv

sqlalchemy
psycopg2-binary

passlib[bcrypt]
bcrypt==4.0.1
python-jose

pandas
numpy

scikit-learn

torch

faiss-cpu

langchain
langchain-openai
openai
tiktoken
python-multipart
pytest
sentence-transformers
transformers
accelerate
bitsandbytes
einops